Updating control pane and folder structure
All standard viewer related components in /viewer all tour specific components (info/aichat) in /tour and all race specific components will be kept in /race
